// KPICalculation/src/app/model/dhrVerticalKpiReport.model.ts
export class DhrVerticalKpi{
    id?: Number;
    validYear?: string;
    description?: string;
    weightage?: string;
    target?: string;
    targetAchieved?: string;
    period?: string
    role?: string;
    weightageAchieved?: string;
    employeeId?: string;
    verticalId?: Number;
    kpiId?: Number;
}